Convert Card to struct

+++ src/Game/Tableau.cs
@@ -32,7 +32,7 @@ namespace FreeCellSolver.Game
             Create(new[] { 0 }.SelectMany(i => cards
                 .Replace(" ", "")
                 .GroupBy(_ => i++ / 2)
-                .Select(g => (byte)Card.Get(string.Join("", g)).RawValue)
+                .Select(g => Card.Get(string.Join("", g)).RawValue)
             ).ToArray());
 
         internal static Tableau Create(Span<byte> cards)
@@ -71,7 +71,7 @@ namespace FreeCellSolver.Game
         public void Push(Card card)
         {
             Debug.Assert(CanPush(card));
-            _state[Size++] = (byte)card.RawValue;
+            _state[Size++] = card.RawValue;
             SortedSize++;
             Top = card;
             Debug.Assert(SortedSize == CountSorted());
